Handle missing EOLs in read()
Need to add in missing last buffer EOL when appending lines,
and need to add in missing last EOL in read lines when
inserting.

In order to do this, we need to know what EOL char to
insert. Hopefully it'll be consistent (?!), and it could
probably be a config attribute eventually (with would then
suggest additional EOL normalization as features). In the
absense of a config item, and to better adapt to existing
files, the code will assume the most used EOL char is the
expected default.