Spin undo_stack logic out into own module.
Will allow special logic for pushing new vs existing item
to undo stack. In the former case, if the redo stack has
existing items, we need to drain the redo stack (to allow
the user to "undo the undos" (i.e., not lose history).
In the latter case, simply pushing is fine.