Remove apparently redundant cursor member from GapBuffer
Current plan is to have the gap at the cursor at all times,
since that optimizes for character insertion, which should
be the most common case. For a screen editor, where might
resonably be the case that one might navigate around
quite a bit in between text input, it make more sense to
let cursor and gap differ and resolve just in time for
text insertion.