handle_char_typed edge case cursor movement addressed
There are some tricky edge cases around double width characters.
Specifically, while I was handling a double width character not
fitting on the line, and thus wrapping to the next line and leaving
the previous line not quite full, I also needed to handle typing
a char when the cursor is on a double width character (i.e., the first
char after the gap is double width).

1. If the new char results in the char after the gap no longer fitting
   on the line, the ag char needs to be wrapped to the beginning of
   the next line, and the cursor shoud follow.
2. If the cursor is in column 0 on a double width character
   and the typed char fits on the previous line, the cursor shouldn't
   move (but we do need to adjust the last line's index accordingly).